Remove SilasProvider
Some collision with the custom Strategy class means it adds more
complexity than it removes
